media-source/SourceBuffer-appendWindowEnd_invalid.html

<!doctype html>
<html>
<head>
  <meta charset='utf-8'>
  <title>Setting SourceBuffer#appendWindowEnd with new invalid value</title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
</head>
<body>
<div>
    <video id='v' controls autoplay width='320' height='240'></video>
</div>
<div id="log"></div>

<script>
var video = document.querySelector('video');
var mimes = ['video/webm; codecs="vorbis,vp8"', 'video/mp4'];

//check the browser supports the MIME used in this test
function isTypeSupported(mime) {
    if(!MediaSource.isTypeSupported(mime)) {
        this.step(function() {
            assert_unreached("Browser doesn't support the MIME used in this test: " + mime);
        });
        this.done();
        return false;
    }
    return true;
}
function test(mime) {
    async_test(function(t) {
        if(!isTypeSupported.bind(t)(mime)) {
            return;
        }
        var mediaSource = new MediaSource();
        mediaSource.addEventListener('sourceopen', function(e) {
            var sourceBuffer = mediaSource.addSourceBuffer(mime);
            t.step(function() {
                assert_throws({name: 'InvalidAccessError'},
                              function() {
                                  sourceBuffer.appendWindowEnd = Number.NaN;
                              },
                              'The new value is NaN');
                assert_throws({name: 'InvalidAccessError'},
                              function() {
                                  sourceBuffer.appendWindowStart = 1;
                                  sourceBuffer.appendWindowEnd = sourceBuffer.appendWindowStart - 1;
                              },
                              'The new value is less than appendWindowStart');
                assert_throws({name: 'InvalidAccessError'},
                              function() {
                                  sourceBuffer.appendWindowEnd = sourceBuffer.appendWindowStart;
                              },
                              'The new value is equal to appendWindowStart');
            });
            t.done();
        });
        video.src = window.URL.createObjectURL(mediaSource);
    }, 'SourceBuffer#appendWindowEnd (' + mime + ') : If the new value equals NaN, INVALID_ACCESS_ERR exception should be thrown.' +
       'If the new value is less than or equal to appendWindowStart, INVALID_ACCESS_ERR exception should be thrown.');
}
mimes.forEach(function(mime) {
    test(mime);
});
</script>
</body>
</html>
